Fixed the rainbow colors and cleaned up everything
