+++ Ktisis/Interface/Types/KtisisWindow.cs
@@ -36,9 +36,7 @@ public abstract class KtisisWindow : Window {
 #else
 	) : base($"{Ktisis.Locale.Translate(localeWindowName)}{windowId}", flags, forceMainWindow) {
 #endif
-#if TESTING
-		this._windowColor = new();
-#endif
+
 		this._localeWindowName = localeWindowName;
 		this._windowId = windowId;
 		this.RespectCloseHotkey = false;
@@ -86,51 +84,4 @@ public abstract class KtisisWindow : Window {
 		});
 	}
 	
-	// Testing branch stuff
-	#if TESTING
-	private ImRaii.ColorDisposable? _windowColor;
-	private List<ImGuiCol> _properties = new List<ImGuiCol>() {
-		ImGuiCol.TitleBg,
-		ImGuiCol.TitleBgCollapsed,
-		ImGuiCol.TitleBgActive
-	};
-	public override void PreDraw() {
-		
-		foreach (var col in this._properties) 
-			this.PushColor(col);
-
-	}
-	public unsafe void PushColor(ImGuiCol col) {
-		var colVec = ImGui.GetStyleColorVec4(col);
-		Vector4 t = new Vector4(colVec->X, colVec->Y, colVec->Z, colVec->W );
-		Vector4* vec = &t;
-		var M = MathF.Max(MathF.Max(vec->X, vec->Y), vec->Z);
-		var m = MathF.Min(MathF.Min(vec->X, vec->Y), vec->Z);
-		var o = M - m;
-		float H, S;
-		
-		if (M == 0) {
-			S = 0;
-		} else {
-			S = o / M;
-		}
-
-		if(o == 0) {
-			H = 0;
-		}else if (M == vec->X) {
-			H = 60 * (((vec->Y - vec->Z) / o) % 360);
-		}else if (M == vec->Y) {
-			H = 60 * (((vec->Z - vec->X) / o) + 2);
-		} else {
-			H = 60 * (((vec->X - vec->Y) / o) + 4);
-		}
-		H += .5f;
-		ImGui.ColorConvertHSVtoRGB(H, S, M, &vec->X, &vec->Y, &vec->Z);
-
-		this._windowColor!.Push(col, *vec);
-	} 
-	public override void PostDraw() {
-		this._windowColor?.Dispose();
-	}
-	#endif
 }
